requirements.txt
azure-common==1.1.28
azure-core==1.35.0
azure-identity==1.24.0
azure-mgmt-core==1.6.0
azure-mgmt-network==29.0.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.6
idna==3.10
isodate==0.7.2
msal==1.33.0
msal-extensions==1.3.1
pycparser==2.22
PyJWT==2.10.1
requests==2.32.4
six==1.17.0
typing_extensions==4.14.1
urllib3==2.5.0
